Fix non-fitting integer type for comparison.
Casting from unsigned to signed integer is safe here.

+++ lib/3rd_party_adapters/SerialInputStreamBuffer.cpp
@@ -18,7 +18,7 @@ SerialInputStreamBuffer::int_type SerialInputStreamBuffer::underflow()
     }
     static const auto bufferLength = std::distance(buffer_begin, buffer_end);
     static const auto bufferSize = bufferLength / sizeof(*buffer_begin);
-    const auto readBytes = Serial.readBytes(buffer_begin, std::min(bufferSize, availableBytes));
+    const auto readBytes = Serial.readBytes(buffer_begin, std::min(static_cast<int>(bufferSize), availableBytes));
     if (readBytes <= 0)
     {
         return traits_type::eof();
